+++ user_channel/Makefile.am
@@ -24,6 +24,7 @@ libuserlayer_la_LDFLAGS = \
 	$(PHOSPHOR_DBUS_INTERFACES_LIBS) \
 	-lstdc++fs \
 	$(CRYPTO_LIBS) \
+	-lpam_misc \
 	-version-info 0:0:0 -shared
 libuserlayer_la_CXXFLAGS = \
 	-I$(top_srcdir) \
